Get the SSL Certificate from an environment variable
Ensure the connection is closed so that it may be reused

WIP fixing and updating version of Go to provide access to http.NewRequestWithContext

CAPT-71 reuse http connection

Configure the HTTP connection to ensure the connection is reused.

Provide test for reusing connections. Limitation of both Go 1.12 and my knowledge means this is not feasible to test so a parallel test is provide which mimics the behaviour of the Audit function in httpSink.go.

It is limited by the fact the types for client and clienttrace are different and therefore a logging http client can't be offered to Audit under test conditions.

Instead this test is verifying that http connections can be reused.

CAPT-71 reuse http connection

Configure the HTTP connection to ensure the connection is reused.

Provide test for reusing connections. Limitation of both Go 1.12 and my knowledge means this is not feasible to test so a parallel test is provide which mimics the behaviour of the Audit function in httpSink.go.

It is limited by the fact the types for client and clienttrace are different and therefore a logging http client can't be offered to Audit under test conditions.

Instead this test is verifying that http connections can be reused.

CAPT-71 include test helpers and test certs

CAPT-71 include test helpers and test certs